+++ src/Pages/DashBoard/Dashboard.tsx
@@ -1,85 +1,43 @@
 import React from 'react';
 import Sidebar from "../../stories/SideBar/sideBar";
-import style from './Dashboard.module.css';
+import style from '../OverView/OverviewPage.module.css';
 import {useNavigate} from "react-router-dom";
 import ResponsiveTable from "./playGround";
-import Profile from "../../Components/DashBoard/Profile";
+import Profile from "../Profile/Profile";
+import OverviewPage from "../OverView/OverviewPage";
 
 
 const Dashboard = () => {
-    const [header, setHeader] = React.useState('Overview');
     const [isSidebarOpen, setIsSidebarOpen] = React.useState<boolean>(false);
+    const [currentPage, setCurrentPage] = React.useState<string>(
+        localStorage.getItem('currentPage') || 'Overview'
+    );
 
-    const handleHeader = (value: string) => {
-        setHeader(value);
-    }
 
     const getSidebarState = (x: boolean): boolean => {
         setIsSidebarOpen(x);
-        // console.log(x, 'dashBoard');
         return x;
     }
-    //
-    // console.log(isSidebarOpen, 'dashBoardisSidebarOpen');
+
+    const getPage = (x: string): string => {
+        setCurrentPage(x);
+     localStorage.setItem('currentPage', x);
+        return x;
+    }
+
 
 
     return (
         <div>
-            <Sidebar logo={'/'} profile={'/saved'} overview={'/dashboard'} getSidebarState={getSidebarState}/>
+            <Sidebar logo={'/'} getSidebarState={getSidebarState} getPage={getPage}/>
             <div className={`${style.container} ${isSidebarOpen ? style.shifted : ''}`}>
-                <div className={style.headerBtnCont}>
-                    <div className={style.headerBtn}>
-                        <div className={style.Btn}>
-                            <div className={style.header} style={{color: header === 'Overview' ? 'black' : '', borderBottom: header === 'Overview' ? 'solid 2px black' : ''}} onClick={() => handleHeader(('Overview'))}>Overview</div>
-                            <div className={style.header} style={{color: header === 'Job' ? 'black' : '', borderBottom: header === 'Job' ? 'solid 2px black' : ''}} onClick={() => handleHeader(('Job'))}>Job Applications</div>
-                        </div>
-                        <div className={style.timeFrame}>
-                            Last 30 days <img src="https://res.cloudinary.com/do5wu6ikf/image/upload/v1725753843/Reev/Icon_Stroke_d2hmut.svg" alt="arrowDown"/>
-                        </div>
-                    </div>
-                    <br/>
-                    <div>
-                        {header === 'Overview' &&
-                            <div className={style.overview}>
-                                <div className={style.EstiActiveCont}>
-                                    <div className={style.EstimateCont}>
-                                        <div className={style.Estimate}>
-                                            <img src="https://res.cloudinary.com/do5wu6ikf/image/upload/v1725752189/Reev/Frame_1171275861_qbbiiv.svg" alt="earn"/>
-                                            <div className={style.earnText}>Total money earned</div>
-                                            <div className={style.earnValue}>$53,00989</div>
-                                            <div className={style.increaseText}><img src="https://res.cloudinary.com/do5wu6ikf/image/upload/v1725752188/Reev/Frame_1171275867_mjlidk.svg" alt="increase"/>12% increase from last month</div>
-                                        </div>
-                                        <div className={style.Estimate} style={{background: 'rgba(240, 194, 116, 0.15)'}}>
-                                            <img src="https://res.cloudinary.com/do5wu6ikf/image/upload/v1725752188/Reev/client_ibnny6.svg" alt="earn"/>
-                                            <div className={style.earnText}>Clients</div>
-                                            <div className={style.earnValue}>101 <span>/120</span></div>
-                                            <div className={style.increaseText}>
-                                                <img src="https://res.cloudinary.com/do5wu6ikf/image/upload/v1725752188/Reev/Frame_1171275867_mjlidk.svg" alt="increase"/>2% increase from last month</div>
-                                        </div>
-                                    </div>
-                                    <div className={style.ActiveProjectCont}>
-                                        <ResponsiveTable/>
-                                    </div>
+                {currentPage === 'Overview' &&
+                    <OverviewPage/>
+                }
 
-                                </div>
-                                <div className={style.ProfileCont}>
-                                    <Profile/>
-                                </div>
-                            </div>
-                        }
-
-                        {header === 'Job' && <div className={style.Job}>
-                            In progress......
-                        </div>
-                        }
-
-                    </div>
-
-                </div>
-
-                <div>
-                    {/*edit profile component*/}
-                </div>
+                {currentPage === 'Profile' &&
+                    <Profile />
+                }
             </div>
         </div>
     );
