+++ src/Pages/Profile/Profile.tsx
@@ -33,7 +33,6 @@ const Profile = () => {
           content="https://res.cloudinary.com/do5wu6ikf/image/upload/v1715619760/Reev/reev_nu0qvs.png"
         />
       </Helmet>
-      <Sidebar logo={'/'} profile={'/saved'} overview={'/'}/>
       <section className={style.container}>
         <div className={style.holder}>
           <div className={style.avatar_container}>
