+++ src/Pages/Profile/profileSave.tsx
@@ -212,7 +212,7 @@ export default function ProfileSave() {
           content="https://res.cloudinary.com/do5wu6ikf/image/upload/v1715619760/Reev/reev_nu0qvs.png"
         />
       </Helmet>
-      <Sidebar logo={'/'} profile={'/saved'} overview={'/'}/>
+      <Sidebar logo={'/'}  />
 
 
       <section className={style.container}>
